Fixed cycle issue
