+++ webapp/src/main/java/com/box/l10n/mojito/service/assetcontent/AssetContentCleanupJob.java
@@ -2,23 +2,16 @@ package com.box.l10n.mojito.service.assetcontent;
 
 import org.quartz.DisallowConcurrentExecution;
 import org.quartz.Job;
-import org.quartz.JobDetail;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.scheduling.quartz.CronTriggerFactoryBean;
-import org.springframework.scheduling.quartz.JobDetailFactoryBean;
 import org.springframework.stereotype.Component;
 
 @Profile("!disablescheduling")
-@Configuration
 @Component
 @DisallowConcurrentExecution
 @ConditionalOnProperty(value = "l10n.asset-content.cleanup-job.enabled", havingValue = "true")
@@ -35,23 +28,5 @@ public class AssetContentCleanupJob implements Job {
     this.assetContentService.cleanAssetContentData(
         this.configurationProperties.getRetentionPeriod(),
         this.configurationProperties.getBatchSize());
-  }
-
-  @Bean(name = "jobDetailAssetContentCleanup")
-  public JobDetailFactoryBean jobDetailAssetContentCleanup() {
-    JobDetailFactoryBean jobDetailFactory = new JobDetailFactoryBean();
-    jobDetailFactory.setJobClass(AssetContentCleanupJob.class);
-    jobDetailFactory.setDescription("Delete old Asset Content data");
-    jobDetailFactory.setDurability(true);
-    return jobDetailFactory;
-  }
-
-  @Bean
-  public CronTriggerFactoryBean triggerAssetContentCleanup(
-      @Qualifier("jobDetailAssetContentCleanup") JobDetail job) {
-    CronTriggerFactoryBean trigger = new CronTriggerFactoryBean();
-    trigger.setJobDetail(job);
-    trigger.setCronExpression(this.configurationProperties.getCron());
-    return trigger;
   }
 }
